Move queryset manipulation logic to managers.py
